// models/userModel.ts
export interface UserModel {
    name: string;
    email: string;
    password: string;
    title: string;
    dayOfBirth: string;
    monthOfBirth: string;
    yearOfBirth: string;
    subscribeToNewsletter: boolean;
    receiveSpecialOffers: boolean;
    firstName: string;
    lastName: string;
    company: string;
    address: string;
    address2: string;
    country: string;
    state: string;
    city: string;
    zipcode: string;
    mobileNumber: string;
}